Steelhead: Fix VCXO control after the latest down-integrate.
Power management code recently added by TI completely broke our ability to use
their timer libraries to control a GP timer in PWM mode in order to control our
VCXO.  This change basically bypasses their libraries and goes directly to the
registers, hopefully insulating us from any future changes to the timer libs.

+++ include/linux/aah_localtime.h
@@ -46,7 +46,7 @@ struct aah_localtime_platform_data {
 	 * should attempt to speed up its counter.  The correction parameter is
 	 * always absolute, not cumulative and is expressed in ppm.
 	 */
-	 void (*set_counter_slew_rate)(s32 correction);
+	 void (*set_counter_slew_rate)(s16 correction);
 
 #ifdef CONFIG_AAH_TIMESYNC_DEBUG
 	/**********************************************************************
@@ -80,7 +80,7 @@ struct aah_localtime_platform_data {
 
 #define AAHLT_IOCTL_LOCALTIME_GET       _IOR(AAH_LOCALTIME_MAGIC, 1, __s64)
 #define AAHLT_IOCTL_LOCALTIME_GETFREQ   _IOR(AAH_LOCALTIME_MAGIC, 2, __u64)
-#define AAHLT_IOCTL_LOCALTIME_SET_SLEW  _IOW(AAH_LOCALTIME_MAGIC, 3, __s32)
+#define AAHLT_IOCTL_LOCALTIME_SET_SLEW  _IOW(AAH_LOCALTIME_MAGIC, 3, __s16)
 
 #ifdef CONFIG_AAH_TIMESYNC_DEBUG
